Fix Compile Errors and Warnings
- Update parking_lot to 0.10
- Update glutin to 0.23
- Remove an unnecessary cast
- into_iter() -> iter()

+++ src/backend/gl/src/device.rs
@@ -1942,8 +1942,8 @@ impl d::Device<B> for Device {
             extent: config.extent,
             context: {
                 surface.context().resize(glutin::dpi::PhysicalSize::new(
-                    config.extent.width as f64,
-                    config.extent.height as f64,
+                    config.extent.width,
+                    config.extent.height,
                 ));
                 surface.context.clone()
             },
